use useCallback to set resizeObserver when node is not null.
