+++ cmp/chart/Chart.js
@@ -5,7 +5,7 @@
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 import PT from 'prop-types';
-import {useRef} from 'react';
+import {useCallback} from 'react';
 import {assign, castArray, clone, isEqual, merge, omit} from 'lodash';
 import {bindable, runInAction} from '@xh/hoist/mobx';
 import {Highcharts} from '@xh/hoist/kit/highcharts';
@@ -13,7 +13,7 @@ import {Highcharts} from '@xh/hoist/kit/highcharts';
 import {XH, hoistCmp, uses, useLocalModel, HoistModel} from '@xh/hoist/core';
 import {div, box} from '@xh/hoist/cmp/layout';
 import {createObservableRef} from '@xh/hoist/utils/react';
-import {getLayoutProps, useOnResize, useOnVisibleChange} from '@xh/hoist/utils/react';
+import {getLayoutProps, useOnVisibleChange, onResize} from '@xh/hoist/utils/react';
 
 import {LightTheme} from './theme/Light';
 import {DarkTheme} from './theme/Dark';
@@ -34,8 +34,12 @@ export const [Chart, chart] = hoistCmp.withFactory({
 
     render({model, className, aspectRatio, ...props}) {
         const impl = useLocalModel(() => new LocalModel(model)),
-            ref = useRef(null);
-        useOnResize(impl.onResize, {ref});
+            ref = useCallback(node => {
+                if (node !== null) {
+                    onResize(impl.onResize, {ref: {current: node}});
+                }
+            }, [impl.onResize]);
+
         useOnVisibleChange(impl.onVisibleChange, {ref});
 
         impl.setAspectRatio(aspectRatio);
